refactor: move workflow command handlers to commands/workflow.py
Extracted all workflow_app and workflow_catalog_app command handlers
from __init__.py into src/specify_cli/commands/workflow.py using the
sub-typer pattern consistent with preset.py and extension.py.

- Replaced 660 lines of inline handlers with a 3-line import + add_typer
- Added module-level constants _SPECIFY_DIR and _WORKFLOWS_SUBDIR
- All handlers carry -> None return type annotations
- Lazy inline imports used for WorkflowEngine, WorkflowRegistry, etc.
- No circular imports; __init__.py now ~340 lines (was 995)
